+++ src/models/cbqlApiQuery.ts
@@ -1,7 +1,7 @@
 /**
  * Defines structure of the payload to query CB Entities with generically
  */
-export interface CodeBeamerItemsQuery {
+export interface CbqlApiQuery {
 	page: number;
 	pageSize: number;
 	queryString: string;